!539 remove linear silent check * remove import os * remove import os * remove embedding and layer norm hook * remove linear silent check and add layer norm check * change loss scale to 1 with fp32 optimizer * move silent check to optimizer

+++ ascendspeed/model/rms_norm.py
@@ -31,7 +31,7 @@ class RMSNorm(torch.nn.Module):
             from torch_npu.hook_module.hook_module import HOOKModule
 
             def norm_hook(grad):
-                HOOKModule.layernorm_list.append(torch.norm(grad))
+                HOOKModule.layernorm_list.append(torch.norm(grad.float()))
 
             self.weight.register_hook(norm_hook)
